Module spidev::spidevioctl

Structs

SpidevTransfer
Representation of a spidev transfer that is shared with external users

spi_ioc_transfer
Structure that is used when performing communication with the kernel.

Functions

get_bits_per_word

get_lsb_first

get_max_speed_hz

get_mode

set_bits_per_word

set_lsb_first

set_max_speed_hz

set_mode

transfer

transfer_multiple