Type Alias spidev::spidevioctl::SpidevTransfer

pub type SpidevTransfer<'a, 'b> = spi_ioc_transfer<'a, 'b>;

Representation of a spidev transfer that is shared with external users

Aliased Type

struct SpidevTransfer<'a, 'b> {
    pub speed_hz: u32,
    pub delay_usecs: u16,
    pub bits_per_word: u8,
    pub cs_change: u8,
    pub pad: u32,
    /* private fields */
}

Fields
speed_hz: u32delay_usecs: u16bits_per_word: u8cs_change: u8pad: u32
Implementations

impl<'a, 'b> spi_ioc_transfer<'a, 'b>

pub fn read(buff: &'b mut [u8]) -> Self

Create a read transfer

pub fn write(buff: &'a [u8]) -> Self

Create a write transfer

pub fn read_write(tx_buf: &'a [u8], rx_buf: &'b mut [u8]) -> Self

Create a read/write transfer. Note that the tx_buf and rx_buf must be the same length.

pub fn delay(microseconds: u16) -> Self

Create a delay transfer of a number of microseconds